refactor: organise florence2 JS logic into folders
