+++ nodes/web/helpers/florence2/main.js
@@ -1,0 +1,3 @@
+import { updateInputsOutputs } from "./utils.js";
+
+export { updateInputsOutputs };
